Refactor transfer to add SuggestedTransferCases
Jira ticket: CAMS-338
