+++ user-interface/src/data-verification/transfer/SuggestedTransferCases.test.tsx
@@ -1,0 +1,5 @@
+import { describe } from 'vitest';
+
+describe('SuggestedTransferCases component', () => {
+  test('should ', () => {});
+});
